requirements.txt
alembic==1.3.2
asn1crypto==1.2.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
bottle==0.12.18
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
dataset==1.1.2
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
Janome==0.3.10
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
kiwisolver==1.1.0
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.2
numpy==1.17.4
odfpy==1.4.0
olefile==0.46
openpyxl==2.4.11
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.6.0
records==0.5.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
tablib==0.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
vaderSentiment==3.2.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0